backend/pyproject.toml
[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git|\.hg|\.mypy_cache|\.tox|\.venv|venv|
    _build|buck-out|build|dist
)/
'''

[tool.bandit]
exclude_dirs = ["tests", "venv"]

[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.0" # Upgrading FastAPI to a version compatible with Starlette >=0.47.2
starlette = ">=0.47.2,<0.48.0" # Keeping Starlette version for security fixes
uvicorn = "^0.30.1"
gunicorn = "^23.0.0" # Upgraded to fix vulnerability
sqlalchemy = "^2.0.30"
passlib = "^1.7.4"
python-multipart = "^0.0.18" # Upgraded to fix vulnerability
bcrypt = "^3.2.0"
PyJWT = "^2.10.1"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
python-json-logger = "^2.0.7"
prometheus_client = "^0.20.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
bandit = "^1.7.9"
pytest = "^8.2.2"
httpx = "^0.28.1"
mypy = "^1.18.2"
types-requests = "^2.32.4.20250913"
types-passlib = "^1.7.7.20250602"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
